requirements.txt
sentence-transformers==2.2.2
scikit-learn==1.3.0
networkx==3.1
matplotlib==3.7.1
pyyaml==6.0
numpy==1.24.3